requirements.txt
streamlit
google.generativeai
python.dotenv
streamlit-chat
langchain
langchain-community